LayoutTests/picture/working/picture-img-alt-tag-only-whitespace.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div tabindex="0" role="group" id="images">
<picture id="firstimage" alt=" " src="../resources/cake.png"></picture><br>
<picture alt="	" src="../resources/cake.png"></picture><br>
<picture alt="   Image " src="../resources/cake.png"></picture><br>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that images with alt tags that only have white space are ignored.");

    if (window.accessibilityController) {

        document.getElementById("images").focus();
        var imagesGroup = accessibilityController.focusedElement;
        shouldBe("imagesGroup.childrenCount", "1");
        console.log("Image description: " + imagesGroup.childAtIndex(0).description);
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
